Update Map converter section to a dialog on the homepage
